+++ data-raw/nsidc.R
@@ -1,11 +1,12 @@
 ## code to prepare `DATASET` dataset goes here
 
 #nsidc_south_files <- raadfiles::nsidc_daily_files()
-nsidc_south_files <- raadtools::icefiles()
-fulldates <- seq(min(nsidc_south_files$date), max(nsidc_south_files$date), by = "1 day")
+nsidc_south_sources <- dplyr::transmute(raadtools::icefiles(hemisphere = "south"),
+                                     date, url = sprintf("ftp://sidads%s", gsub(".*sidads", "", fullname)))
 
-ftp <- sprintf("ftp://sidads%s", gsub(".*sidads", "", nsidc_south_files$fullname))
-idx <- findInterval(as.integer(fulldates), as.integer(nsidc_south_files$date))
-nsidc_south_files <- tibble::tibble(date = fulldates, url = ftp[idx], miss = duplicated(idx))
+nsidc_north_sources <- dplyr::transmute(raadtools::icefiles(hemisphere = "north"),
+                                        date, url = sprintf("ftp://sidads%s", gsub(".*sidads", "", fullname)))
 
-usethis::use_data(nsidc_south_files, overwrite = TRUE)
+
+usethis::use_data(nsidc_south_sources, overwrite = TRUE)
+usethis::use_data(nsidc_north_sources, overwrite = TRUE)